requirements.txt
# requirements.txt

streamlit
langchain
langchain_community
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv